requirements.txt
PyPDF2
pdfminer.six 
beautifulsoup4
markdown
markdownify